Simplify command line interface
