+++ spec/integration/cli/build_spec.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-RSpec.describe '`spandx index build` command', type: :cli do
-  it 'executes `spandx index help build` command successfully' do
-    output = `spandx index help build`
+RSpec.describe '`spandx build` command', type: :cli do
+  it 'executes `spandx help build` command successfully' do
+    output = `spandx help build`
     expected_output = <<~OUT
       Usage:
-        spandx index build
+        spandx build
 
       Options:
         -h, [--help], [--no-help]    # Display usage information
